Packet.pm: apply memoization to get_records()
This function may not seem to be very computationally expensive, but
it’s called very very often. Memoizing it gives a noticable speed boost
of about 2 or 3 percent.

It’s a bit of a sledgehammer approach and it might be better to
re-engineer it altogether in order to make better space-time tradeoffs.
At least, it goes to show that memoization can be interesting for
functions like these.