Emit TEST_CASE_{START,END} without repeating code
Test methods in Zonemaster-Engine all suffer from the same boilerplate
code repeated all over the place. Especially having to remember to
generate TEST_CASE_END messages at every exit point of a test case is a
recipe for mistakes.

Isn’t there a better way? How about making sure that we don’t need to
repeat ourselves in the code for test cases? Wouldn’t it be nice if we
could leverage Perl to do something on entry and on exit without
repeating ourselves?

Fortunately, Perl offers a feature called subroutine attributes, which
we can use for some brutal metaprogramming. Attributes in general are a
bit of an obscure feature (also because "perldoc attributes" merely
serves as a reference, not as a tutorial), but their semantics are
well-defined when used on subroutines. They are also extensively used in
Perl Catalyst, a Web framework that is just about to celebrate its 20th
birthday this year (in 2025).

This commit starts by marking each test case method with the “TestCase”
attribute, like so:

    sub example01 : TestCase {
        ...;
    }

At compilation time, Perl invokes a function named
MODIFY_CODE_ATTRIBUTES in the context of the subroutine definition. The
implementation of MODIFY_CODE_ATTRIBUTES lives in a new module,
Zonemaster::Engine::TestCase, which, among other things, silently exports
that function into the namespace of any package having "use
Zonemaster::Engine::TestCase".

Then, the MODIFY_CODE_ATTRIBUTES function finds the name of the
function, the name of the module that defines it, wraps the body of the
subroutine in a subroutine that also emits TEST_CASE_START before and
TEST_CASE_END after execution, and finally redefines the subroutine to
call that function instead of the original.

This way, there is a visual marker (the “TestCase” attribute) that tells
the reader that these test case functions are slightly special, and we
no longer need to worry that each code path emits TEST_CASE_START and
TEST_CASE_END ourselves.