Inject MODIFY_CODE_ATTRIBUTES in caller modules
As explained in the previous commit, MODIFY_CODE_ATTRIBUTES is the
function that’s called when a subroutine is defined with at least one
attribute.

But it’s the package defining the test cases that needs to have this
function. In Zonemaster::Engine::TestCase, however, we really don’t need
to have such a symbol in our namespace. Rewriting the module so as to
modify the caller package’s symbol table directly states our intent more
clearly, and lets us define more subroutines that way if we want.